WIP: 2021 rewrite

+++ asu/build.py
@@ -14,7 +14,43 @@ log = logging.getLogger("rq.worker")
 log.setLevel(logging.DEBUG)
 
 
-def build(request: dict):
+class PackageSelectionError(Exception):
+    pass
+
+
+class ImageBuildError(Exception):
+    pass
+
+
+class JSONMissingError(Exception):
+    pass
+
+
+class JSONMissingProfileError(Exception):
+    pass
+
+
+class StorePathMissingError(Exception):
+    pass
+
+
+class ChecksumMissingError(Exception):
+    pass
+
+
+class BadChecksumError(Exception):
+    pass
+
+
+class BadSignatureError(Exception):
+    pass
+
+
+class ExtractArchiveError(Exception):
+    pass
+
+
+def build(req: dict):
     """Build image request and setup ImageBuilders automatically
 
     The `request` dict contains properties of the requested image.
@@ -23,25 +59,19 @@ def build(request: dict):
         request (dict): Contains all properties of requested image
     """
 
-    if "packages" in request:
-        assert isinstance(
-            request["packages"], set
-        ), "packages must be type set not list"
-    else:
-        request["packages"] = set()
-
-    assert (request["store_path"]).is_dir(), "store_path must be existing directory"
+    if not req["store_path"].is_dir():
+        raise StorePathMissingError()
 
     job = get_current_job()
 
-    log.debug(f"Building {request}")
-    cache = (request["cache_path"] / request["version"] / request["target"]).parent
-    target, subtarget = request["target"].split("/")
+    log.debug(f"Building {req}")
+    cache = (req["cache_path"] / req["version"] / req["target"]).parent
+    target, subtarget = req["target"].split("/")
     sums_file = Path(cache / f"{subtarget}_sums")
     sig_file = Path(cache / f"{subtarget}_sums.sig")
 
     def setup_ib():
-        """Setup ImageBuilder based on `request`
+        """Setup ImageBuilder based on `req`
 
         This function downloads and verifies the ImageBuilder archive. Existing
         setups are automatically updated if newer version are available
@@ -54,26 +84,24 @@ def build(request: dict):
         download_file("sha256sums.sig", sig_file)
         download_file("sha256sums", sums_file)
 
-        assert verify_usign(
-            sig_file, sums_file, request["version_data"]["pubkey"]
-        ), "Bad signature for cheksums"
+        if not verify_usign(sig_file, sums_file, req["branch_data"]["pubkey"]):
+            raise BadSignatureError()
 
-        # openwrt-imagebuilder-ath79-generic.Linux-x86_64.tar.xz
         ib_search = re.search(
             r"^(.{64}) \*(openwrt-imagebuilder-.+?\.Linux-x86_64\.tar\.xz)$",
             sums_file.read_text(),
             re.MULTILINE,
         )
 
-        assert ib_search, "No ImageBuilder in checksums found"
+        if not ib_search:
+            raise ChecksumMissingError()
 
         ib_hash, ib_archive = ib_search.groups()
 
         download_file(ib_archive)
 
-        assert ib_hash == get_file_hash(
-            cache / ib_archive
-        ), "Wrong ImageBuilder archive checksum"
+        if ib_hash != get_file_hash(cache / ib_archive):
+            raise BadChecksumError()
 
         (cache / subtarget).mkdir(parents=True, exist_ok=True)
         extract_archive = subprocess.run(
@@ -81,21 +109,29 @@ def build(request: dict):
             cwd=cache,
         )
 
-        assert not extract_archive.returncode, "Extracting ImageBuilder archive failed"
+        if extract_archive.returncode:
+            raise ExtractArchiveError()
 
         log.debug(f"Extracted TAR {ib_archive}")
 
         (cache / ib_archive).unlink()
 
-        extra_repos = request["version_data"].get("extra_repos")
+        repos_path = cache / subtarget / "repositories.conf"
+        repos = repos_path.read_text()
+
+        # speed up downloads with HTTP and CDN
+        repos = repos.replace("https://downloads.openwrt.org", req["upstream_url"])
+        repos = repos.replace("http://downloads.openwrt.org", req["upstream_url"])
+        repos = repos.replace("https", "http")
+
+        extra_repos = req["branch_data"].get("extra_repos")
         if extra_repos:
             log.debug("Found extra repos")
-            repos_path = cache / subtarget / "repositories.conf"
-            repos = repos_path.read_text()
             for name, repo in extra_repos.items():
                 repos += f"\nsrc/gz {name} {repo}"
-            repos_path.write_text(repos)
-            log.debug(f"Repos:\n{repos}")
+
+        repos_path.write_text(repos)
+        log.debug(f"Repos:\n{repos}")
 
     def download_file(filename: str, dest: str = None):
         """Download file from upstream target path
@@ -109,11 +145,11 @@ def build(request: dict):
         """
         log.debug(f"Downloading {filename}")
         urllib.request.urlretrieve(
-            request["upstream_url"]
+            req["upstream_url"]
             + "/"
-            + request["version_data"]["path"]
+            + req["branch_data"]["path"].format(version=req["version"])
             + "/targets/"
-            + request["target"]
+            + req["target"]
             + "/"
             + filename,
             dest or (cache / filename),
@@ -124,11 +160,11 @@ def build(request: dict):
     stamp_file = cache / f"{subtarget}_stamp"
 
     sig_file_headers = urllib.request.urlopen(
-        request["upstream_url"]
+        req["upstream_url"]
         + "/"
-        + request["version_data"]["path"]
+        + req["branch_data"]["path"].format(version=req["version"])
         + "/targets/"
-        + request["target"]
+        + req["target"]
         + "/sha256sums.sig"
     ).info()
     log.debug(f"sig_file_headers: \n{sig_file_headers}")
@@ -148,7 +184,7 @@ def build(request: dict):
 
     stamp_file.write_text(origin_modified)
 
-    if request.get("diff_packages", False) and request.get("packages"):
+    if req.get("diff_packages", False) and req.get("packages"):
         info_run = subprocess.run(
             ["make", "info"], text=True, capture_output=True, cwd=cache / subtarget
         )
@@ -157,33 +193,37 @@ def build(request: dict):
         )
         profile_packages = set(
             re.search(
-                r"{}:\n    .+\n    Packages: (.+?)\n".format(request["profile"]),
+                r"{}:\n    .+\n    Packages: (.*?)\n".format(req["profile"]),
                 info_run.stdout,
                 re.MULTILINE,
             )
             .group(1)
             .split()
         )
-        remove_packages = (default_packages | profile_packages) - request["packages"]
-        request["packages"] = request["packages"] | set(
-            map(lambda p: f"-{p}", remove_packages)
-        )
+        remove_packages = (default_packages | profile_packages) - req["packages"]
+        req["packages"] = req["packages"] | set(map(lambda p: f"-{p}", remove_packages))
 
     manifest_run = subprocess.run(
         [
             "make",
             "manifest",
-            f"PROFILE={request['profile']}",
-            f"PACKAGES={' '.join(request['packages'])}",
+            f"PROFILE={req['profile']}",
+            f"PACKAGES={' '.join(req['packages'])}",
         ],
         text=True,
-        capture_output=True,
         cwd=cache / subtarget,
+        capture_output=True,
     )
 
     if manifest_run.returncode:
-        log.error(f"Manifest stdout {manifest_run.stdout}")
-        log.error(f"Manifest stderr {manifest_run.stderr}")
+        if "Package size mismatch" in manifest_run.stderr:
+            rmtree(cache / subtarget)
+            return build(req)
+        else:
+            job.meta["stdout"] = manifest_run.stdout
+            job.meta["stderr"] = manifest_run.stderr
+            job.save_meta()
+            raise PackageSelectionError()
 
     manifest = dict(map(lambda pv: pv.split(" - "), manifest_run.stdout.splitlines()))
 
@@ -194,58 +234,47 @@ def build(request: dict):
     packages_hash = get_packages_hash(manifest_packages)
     log.debug(f"Packages Hash {packages_hash}")
 
-    bin_dir = (
-        Path(request["version"])
-        / request["target"]
-        / request["profile"]
-        / packages_hash
-    )
+    bin_dir = Path(req["version"]) / req["target"] / req["profile"] / packages_hash
 
-    (request["store_path"] / bin_dir).mkdir(parents=True, exist_ok=True)
+    (req["store_path"] / bin_dir).mkdir(parents=True, exist_ok=True)
 
     image_build = subprocess.run(
         [
             "make",
             "image",
-            f"PROFILE={request['profile']}",
-            f"PACKAGES={' '.join(request['packages'])}",
+            f"PROFILE={req['profile']}",
+            f"PACKAGES={' '.join(req['packages'])}",
             f"EXTRA_IMAGE_NAME={packages_hash}",
-            f"BIN_DIR={request['store_path'] / bin_dir}",
+            f"BIN_DIR={req['store_path'] / bin_dir}",
         ],
         text=True,
-        capture_output=True,
         cwd=cache / subtarget,
-    )
-
-    (request["store_path"] / bin_dir / "buildlog.txt").write_text(
-        f"### STDOUT\n\n{image_build.stdout}\n\n### STDERR\n\n{image_build.stderr}"
+        capture_output=True,
     )
 
     # check if running as job or within pytest
     if job:
+        job.meta["stdout"] = image_build.stdout
+        job.meta["stderr"] = image_build.stderr
         job.meta["bin_dir"] = str(bin_dir)
-        job.meta["buildlog"] = True
         job.save_meta()
 
     if image_build.returncode:
-        log.error(f"Build stdout {image_build.stdout}")
-        log.error(f"Build stderr {image_build.stderr}")
+        raise ImageBuildError()
 
-    assert not image_build.returncode, "ImageBuilder failed"
+    json_file = Path(req["store_path"] / bin_dir / "profiles.json")
 
-    json_file = Path(request["store_path"] / bin_dir / "profiles.json")
-
-    assert json_file.is_file(), "Image built but no profiles.json file created"
+    if not json_file.is_file():
+        raise JSONMissingError()
 
     json_content = json.loads(json_file.read_text())
 
-    assert (
-        request["profile"] in json_content["profiles"]
-    ), "Requested profile not in created profiles.json"
+    if req["profile"] not in json_content["profiles"]:
+        raise JSONMissingProfileError()
 
     json_content.update({"manifest": manifest})
-    json_content.update(json_content["profiles"][request["profile"]])
-    json_content["id"] = request["profile"]
+    json_content.update(json_content["profiles"][req["profile"]])
+    json_content["id"] = req["profile"]
     json_content.pop("profiles")
 
     return json_content
